use long for entity ids

+++ api/src/main/java/org/ccci/gto/globalreg/AbstractGlobalRegistryClient.java
@@ -7,7 +7,7 @@ public abstract class AbstractGlobalRegistryClient implements GlobalRegistryClie
     public static final int DEFAULT_PAGE = 1;
 
     @Override
-    public final <T> T getEntity(final Type<T> type, final int id) throws SerializerException, UnauthorizedException {
+    public final <T> T getEntity(final Type<T> type, final long id) throws SerializerException, UnauthorizedException {
         return this.getEntity(type, id, DEFAULT_CREATED_BY);
     }
